impl(lro): `internal` module for internal types
For types that we must make public, but are not intended as public
APIs we are using `internal` somewhere in its module name.
