+++ src/lro/tests/fake/library/builders.rs
@@ -63,7 +63,7 @@ impl CreateResource {
         self,
     ) -> impl google_cloud_lro::Poller<super::model::Resource, super::model::CreateResourceMetadata>
     {
-        type Operation = google_cloud_lro::Operation<
+        type Operation = google_cloud_lro::internal::Operation<
             super::model::Resource,
             super::model::CreateResourceMetadata,
         >;
